# Optimise proposer index calculation in block processing