--- a/beacon_node/beacon_chain/src/beacon_chain.rs
+++ b/beacon_node/beacon_chain/src/beacon_chain.rs
@@ -3405,7 +3405,8 @@ impl<T: BeaconChainTypes> BeaconChain<T> {
         drop(attestation_packing_timer);
 
         let slot = state.slot();
-        let proposer_index = state.get_beacon_proposer_index(state.slot(), &self.spec)? as u64;
+        let proposer_index =
+            state.get_beacon_proposer_index_using_committee_cache(state.slot(), &self.spec)? as u64;
 
         let sync_aggregate = if matches!(&state, BeaconState::Base(_)) {
             None