--- a/consensus/state_processing/src/per_block_processing/block_signature_verifier.rs
+++ b/consensus/state_processing/src/per_block_processing/block_signature_verifier.rs
@@ -74,6 +74,7 @@ where
     state: &'a BeaconState<T>,
     spec: &'a ChainSpec,
     sets: ParallelSignatureSets<'a>,
+    proposer_index: Option<usize>,
 }
 
 #[derive(Default)]
@@ -107,6 +108,30 @@ where
             state,
             spec,
             sets: ParallelSignatureSets::default(),
+            proposer_index: None,
+        }
+    }
+
+    pub fn proposer_index<Payload: ExecPayload<T>>(
+        &mut self,
+        block: &'a SignedBeaconBlock<T, Payload>,
+    ) -> Result<usize> {
+        if let Some(index) = self.proposer_index {
+            Ok(index)
+        } else {
+            let proposer_index = self
+                .state
+                .get_beacon_proposer_index_using_committee_cache(block.slot(), self.spec)?;
+
+            if proposer_index != block.message().proposer_index() as usize {
+                return Err(Error::IncorrectBlockProposer {
+                    block: block.message().proposer_index(),
+                    local_shuffling: proposer_index as u64,
+                });
+            }
+
+            self.proposer_index = Some(proposer_index);
+            Ok(proposer_index)
         }
     }
 
@@ -136,7 +161,8 @@ where
         block: &'a SignedBeaconBlock<T, Payload>,
         block_root: Option<Hash256>,
     ) -> Result<()> {
-        self.include_block_proposal(block, block_root)?;
+        let proposer_index = self.proposer_index(block)?;
+        self.include_block_proposal(block, block_root, proposer_index)?;
         self.include_all_signatures_except_proposal(block)?;
 
         Ok(())
@@ -148,7 +174,8 @@ where
         &mut self,
         block: &'a SignedBeaconBlock<T, Payload>,
     ) -> Result<()> {
-        self.include_randao_reveal(block)?;
+        let proposer_index = self.proposer_index(block)?;
+        self.include_randao_reveal(block, proposer_index)?;
         self.include_proposer_slashings(block)?;
         self.include_attester_slashings(block)?;
         self.include_attestations(block)?;
@@ -164,12 +191,14 @@ where
         &mut self,
         block: &'a SignedBeaconBlock<T, Payload>,
         block_root: Option<Hash256>,
+        proposer_index: usize,
     ) -> Result<()> {
         let set = block_proposal_signature_set(
             self.state,
             self.get_pubkey.clone(),
             block,
             block_root,
+            proposer_index as u64,
             self.spec,
         )?;
         self.sets.push(set);
@@ -180,11 +209,13 @@ where
     pub fn include_randao_reveal<Payload: ExecPayload<T>>(
         &mut self,
         block: &'a SignedBeaconBlock<T, Payload>,
+        proposer_index: usize,
     ) -> Result<()> {
         let set = randao_signature_set(
             self.state,
             self.get_pubkey.clone(),
             block.message(),
+            proposer_index as u64,
             self.spec,
         )?;
         self.sets.push(set);