--- a/consensus/state_processing/src/per_block_processing/tests.rs
+++ b/consensus/state_processing/src/per_block_processing/tests.rs
@@ -640,10 +640,14 @@ async fn valid_insert_attester_slashing() {
     let attester_slashing = harness.make_attester_slashing(vec![1, 2]);
 
     let mut state = harness.get_current_state();
+    let proposer_index = state
+        .get_beacon_proposer_index(state.slot(), &spec)
+        .unwrap();
     let result = process_operations::process_attester_slashings(
         &mut state,
         &[attester_slashing],
         VerifySignatures::True,
+        proposer_index,
         &spec,
     );
 
@@ -660,10 +664,14 @@ async fn invalid_attester_slashing_not_slashable() {
     attester_slashing.attestation_1 = attester_slashing.attestation_2.clone();
 
     let mut state = harness.get_current_state();
+    let proposer_index = state
+        .get_beacon_proposer_index(state.slot(), &spec)
+        .unwrap();
     let result = process_operations::process_attester_slashings(
         &mut state,
         &[attester_slashing],
         VerifySignatures::True,
+        proposer_index,
         &spec,
     );
 
@@ -686,10 +694,14 @@ async fn invalid_attester_slashing_1_invalid() {
     attester_slashing.attestation_1.attesting_indices = VariableList::from(vec![2, 1]);
 
     let mut state = harness.get_current_state();
+    let proposer_index = state
+        .get_beacon_proposer_index(state.slot(), &spec)
+        .unwrap();
     let result = process_operations::process_attester_slashings(
         &mut state,
         &[attester_slashing],
         VerifySignatures::True,
+        proposer_index,
         &spec,
     );
 
@@ -715,10 +727,14 @@ async fn invalid_attester_slashing_2_invalid() {
     attester_slashing.attestation_2.attesting_indices = VariableList::from(vec![2, 1]);
 
     let mut state = harness.get_current_state();
+    let proposer_index = state
+        .get_beacon_proposer_index(state.slot(), &spec)
+        .unwrap();
     let result = process_operations::process_attester_slashings(
         &mut state,
         &[attester_slashing],
         VerifySignatures::True,
+        proposer_index,
         &spec,
     );
 
@@ -741,10 +757,14 @@ async fn valid_insert_proposer_slashing() {
     let harness = get_harness::<MainnetEthSpec>(EPOCH_OFFSET, VALIDATOR_COUNT).await;
     let proposer_slashing = harness.make_proposer_slashing(1);
     let mut state = harness.get_current_state();
+    let proposer_index = state
+        .get_beacon_proposer_index(state.slot(), &spec)
+        .unwrap();
     let result = process_operations::process_proposer_slashings(
         &mut state,
         &[proposer_slashing],
         VerifySignatures::True,
+        proposer_index,
         &spec,
     );
     // Expecting Ok(_) because we inserted a valid proposer slashing
@@ -760,10 +780,14 @@ async fn invalid_proposer_slashing_proposals_identical() {
     proposer_slashing.signed_header_1.message = proposer_slashing.signed_header_2.message.clone();
 
     let mut state = harness.get_current_state();
+    let proposer_index = state
+        .get_beacon_proposer_index(state.slot(), &spec)
+        .unwrap();
     let result = process_operations::process_proposer_slashings(
         &mut state,
         &[proposer_slashing],
         VerifySignatures::True,
+        proposer_index,
         &spec,
     );
 
@@ -787,10 +811,14 @@ async fn invalid_proposer_slashing_proposer_unknown() {
     proposer_slashing.signed_header_2.message.proposer_index = 3_141_592;
 
     let mut state = harness.get_current_state();
+    let proposer_index = state
+        .get_beacon_proposer_index(state.slot(), &spec)
+        .unwrap();
     let result = process_operations::process_proposer_slashings(
         &mut state,
         &[proposer_slashing],
         VerifySignatures::True,
+        proposer_index,
         &spec,
     );
 
@@ -811,10 +839,14 @@ async fn invalid_proposer_slashing_duplicate_slashing() {
 
     let proposer_slashing = harness.make_proposer_slashing(1);
     let mut state = harness.get_current_state();
+    let proposer_index = state
+        .get_beacon_proposer_index(state.slot(), &spec)
+        .unwrap();
     let result_1 = process_operations::process_proposer_slashings(
         &mut state,
         &[proposer_slashing.clone()],
         VerifySignatures::False,
+        proposer_index,
         &spec,
     );
     assert!(result_1.is_ok());
@@ -823,6 +855,7 @@ async fn invalid_proposer_slashing_duplicate_slashing() {
         &mut state,
         &[proposer_slashing],
         VerifySignatures::False,
+        proposer_index,
         &spec,
     );
     // Expecting ProposerNotSlashable because we've already slashed the validator
@@ -842,10 +875,14 @@ async fn invalid_bad_proposal_1_signature() {
     let mut proposer_slashing = harness.make_proposer_slashing(1);
     proposer_slashing.signed_header_1.signature = Signature::empty();
     let mut state = harness.get_current_state();
+    let proposer_index = state
+        .get_beacon_proposer_index(state.slot(), &spec)
+        .unwrap();
     let result = process_operations::process_proposer_slashings(
         &mut state,
         &[proposer_slashing],
         VerifySignatures::True,
+        proposer_index,
         &spec,
     );
 
@@ -866,10 +903,14 @@ async fn invalid_bad_proposal_2_signature() {
     let mut proposer_slashing = harness.make_proposer_slashing(1);
     proposer_slashing.signed_header_2.signature = Signature::empty();
     let mut state = harness.get_current_state();
+    let proposer_index = state
+        .get_beacon_proposer_index(state.slot(), &spec)
+        .unwrap();
     let result = process_operations::process_proposer_slashings(
         &mut state,
         &[proposer_slashing],
         VerifySignatures::True,
+        proposer_index,
         &spec,
     );
 
@@ -891,10 +932,14 @@ async fn invalid_proposer_slashing_proposal_epoch_mismatch() {
     proposer_slashing.signed_header_1.message.slot = Slot::new(0);
     proposer_slashing.signed_header_2.message.slot = Slot::new(128);
     let mut state = harness.get_current_state();
+    let proposer_index = state
+        .get_beacon_proposer_index(state.slot(), &spec)
+        .unwrap();
     let result = process_operations::process_proposer_slashings(
         &mut state,
         &[proposer_slashing],
         VerifySignatures::False,
+        proposer_index,
         &spec,
     );