… consensus/swap_or_not_shuffle/src/compute_shuffled_index.rs
… consensus/swap_or_not_shuffle/src/lib.rs
… consensus/types/src/beacon_state.rs
… consensus/types/src/beacon_state/committee_cache.rs
… lcli/src/transition_blocks.rs
… testing/ef_tests/src/cases/operations.rs
… testing/state_transition_vectors/src/exit.rs