@michaelsproul 2022-06-12
On `tree-states` I've implemented an optimisation that avoids calculating the proposer index at all. Using the existing proposer shuffling cache the index is checked once during block verification and then plumbed through block processing with a `ConsensusContext`. I'm not opposed to merging this change as well/instead, but an alternative would be just to port the `ConsensusContext` from `tree-states` to `unstable`. It also carries the block root through from initial verification