@paulhauner 2022-06-12
Oh yeah cool! We could potentially use this method for building the proposer shuffling cache. We can think more about it next week :)