@paulhauner 2022-11-14
I'm going to close this since it's become much less useful since we're avoiding re-computation of the proposer index. The spiciness of reverse shuffling doesn't seem worth it now.